adda

+++ utils.py
@@ -24,7 +24,9 @@ class Moment:
         encoder.eval()
         datalen = len(dataset)
         if not is_memory:
-            self.features = torch.zeros(datalen, self.config.encoder_output_size, dtype=torch.float32)
+            # self.features = torch.zeros(datalen, self.config.encoder_output_size, dtype=torch.bfloat16)
+            self.features = torch.zeros(datalen, self.config.encoder_output_size)
+
             data_loader = get_data_loader_BERT(self.config, dataset) # shuffle=False
             lbs = []
             for step, (instance, labels, ind) in enumerate(data_loader):
@@ -40,7 +42,9 @@ class Moment:
             self.labels = lbs
         else:
             self.mem_samples = dataset
-            self.mem_features = torch.zeros(datalen, self.config.encoder_output_size, dtype=torch.float32)
+            # self.mem_features = torch.zeros(datalen, self.config.encoder_output_size, dtype=torch.bfloat16)
+            self.mem_features = torch.zeros(datalen, self.config.encoder_output_size)
+
             data_loader = get_data_loader_BERT(self.config, dataset) # shuffle=False
             lbs = []
             for step, (instance, labels, ind) in enumerate(data_loader):
@@ -82,6 +86,8 @@ class Moment:
         num = len(cinds)
         feats = self.mem_features
         centroids = torch.zeros((num, feats.size(1)), dtype=torch.float32, device=feats.device)
+        # centroids = torch.zeros((num, feats.size(1)), dtype=torch.bfloat16, device=feats.device)
+
         for i, c in enumerate(cinds):
             ind = np.where(self.mem_labels.cpu().numpy() == c)[0]
             centroids[i, :] = feats[ind, :].mean(dim=0)
